Add MRI sound playback feature with duration control
Introduced a new "MRI" tab in the UI with a "Play MRI Sound" button
to play a looping MRI sound. Added a `_durationSlider` to set
playback duration (1-60 minutes) and integrated a `_playbackTimer`
to stop playback automatically. Updated control layout and `Resize`
event handler to accommodate the new slider. Refactored
`InitializeMriTab` for setup and adjusted `_startButton` to use the
slider's value. Updated the application icon to `neurotones.ico`
and removed redundant icon initialization code.